requirements.txt
pyrebase
aiohttp
discord.py